+++ Expanded Cheat Menu/Changelog.md
@@ -4,6 +4,8 @@
 ##### Dialogs: Change map, Monitor Info.
 
 ## v8.9 (Unreleased)
+### Added:
+- Cheats>Research>Change Outsource Limit:How many times you can outsource in a row.
 
 ## v8.8 (07 Sep 2018)
 ### Added: